Bugfixes so the code is being run
Had to make sure the sender acct was specified and had gas.
Code is now executing but destination contract code is not being set to
return value of code.
Added an expect line for the code of the created contract (found the
address by running `testeth` with `--statediff` and looking in the
filled test file).